[Fix] 修复 StringUtil 类的 parseFloat() 函数错误的返回值类型, 应为 float 值

+++ src/main/java/com/minecraft/moonlake/util/StringUtil.java
@@ -571,7 +571,7 @@ public class StringUtil {
      * @param str 字符串
      * @return 单精度浮点数, 异常返回 0
      */
-    public static double parseFloat(String str) {
+    public static float parseFloat(String str) {
 
         return parseFloat(str, 0f);
     }
@@ -583,7 +583,7 @@ public class StringUtil {
      * @param def 默认值
      * @return 单精度浮点数, 异常返回默认值
      */
-    public static double parseFloat(String str, float def) {
+    public static float parseFloat(String str, float def) {
 
         return parseFloat((Object) str, def);
     }
@@ -594,7 +594,7 @@ public class StringUtil {
      * @param obj 对象
      * @return 单精度浮点数, 异常返回 0
      */
-    public static double parseFloat(Object obj) {
+    public static float parseFloat(Object obj) {
 
         return parseFloat(obj, 0f);
     }
@@ -606,7 +606,7 @@ public class StringUtil {
      * @param def 默认值
      * @return 单精度浮点数, 异常返回默认值
      */
-    public static double parseFloat(Object obj, float def) {
+    public static float parseFloat(Object obj, float def) {
 
         if(obj instanceof Number)
             return ((Number) obj).floatValue();
